application/requirements.txt
cffi==1.14.0
Click==7.0
cryptography==2.8
dnspython==1.16.0
Flask==1.1.2
Flask-Login==0.5.0
flask-paginate==0.6.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
Pillow==7.0.0
protobuf==3.6.1
pycparser==2.19
PyMySQL==0.9.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
WTForms==2.2.1
